--- src/pages/index.js
@@ -1,17 +1,0 @@
-import React from "react"
-import { Link } from "gatsby"
-
-import Layout from "@layouts/Default"
-import SEO from "@components/SEO"
-
-const IndexPage = () => (
-  <Layout>
-    <SEO title="Home" />
-    <h1>Hi people</h1>
-    <p>Welcome to your new Gatsby site.</p>
-    <p>Now go build something great.</p>
-    <Link to="/page-2/">Go to page 2</Link>
-  </Layout>
-)
-
-export default IndexPage
